cl22clx: Add an option to combine CL2s
Also makes a few functions public to allow doing this from a library.

Example:

    cl22clx --combine --width 96 farrow{1..16}.cl2

+++ src/internal/cl22clx_main.cpp
@@ -1,6 +1,7 @@
 #include <filesystem>
 #include <iostream>
 #include <limits>
+#include <optional>
 #include <string>
 #include <string_view>
 #include <utility>
@@ -21,7 +22,11 @@ Converts CL2 sprite(s) to a CLX file.
 
 Options:
   --output-dir <arg>           Output directory. Default: input file directory.
+  --output-filename <arg>      Output filename. Default: input basename with the ".clx" extension.
+                               With --combine, the default is the basename of the first file without
+                               the trailing digits.
   --width <arg>[,<arg>...]     CL2 sprite frame width(s), comma-separated.
+  --combine                    Combine multiple CL2 files into a single CLX sheet.
   --remove                     Remove the input files.
   -q, --quiet                  Do not log anything.
 )";
@@ -29,7 +34,9 @@ Options:
 struct Options {
 	std::vector<const char *> inputPaths;
 	std::optional<std::string_view> outputDir;
+	std::optional<std::string_view> outputFilename;
 	std::vector<uint16_t> widths;
+	bool combine = false;
 	bool remove = false;
 	bool quiet = false;
 };
@@ -58,11 +65,18 @@ tl::expected<Options, ArgumentError> ParseArguments(int argc, char *argv[])
 			if (!value.has_value())
 				return tl::unexpected { std::move(value).error() };
 			options.outputDir = *value;
+		} else if (arg == "--output-filename") {
+			tl::expected<std::string_view, ArgumentError> value = ParseArgumentValue(state);
+			if (!value.has_value())
+				return tl::unexpected { std::move(value).error() };
+			options.outputFilename = *value;
 		} else if (arg == "--width") {
 			tl::expected<std::vector<uint16_t>, ArgumentError> value = ParseIntListArgument<uint16_t>(state);
 			if (!value.has_value())
 				return tl::unexpected { std::move(value).error() };
 			options.widths = *std::move(value);
+		} else if (arg == "--combine") {
+			options.combine = true;
 		} else if (arg == "--remove") {
 			options.remove = true;
 		} else if (arg == "-q" || arg == "--quiet") {
@@ -77,7 +91,33 @@ tl::expected<Options, ArgumentError> ParseArguments(int argc, char *argv[])
 	    error.has_value()) {
 		return tl::unexpected { *std::move(error) };
 	}
+	if (!options.combine && options.outputFilename.has_value() && options.inputPaths.size() > 1) {
+		return tl::unexpected { ArgumentError {
+			"--output-filename", "Cannot pass more than one input path with --output-filename and without --combine" } };
+	}
+	if (options.combine && options.inputPaths.size() < 2) {
+		return tl::unexpected { ArgumentError {
+			"--combine", "requires at least 2 input files" } };
+	}
+	if (options.widths.empty()) {
+		return tl::unexpected { ArgumentError { "--width", "is required" } };
+	}
 	return options;
+}
+
+std::string DefaultCombinedFilename(const Options &options)
+{
+	std::string outputFilename = std::filesystem::path(options.inputPaths[0]).stem().string();
+	size_t numSuffixLength = 0;
+	while (numSuffixLength < outputFilename.size()) {
+		const char c = outputFilename[outputFilename.size() - numSuffixLength - 1];
+		if (c < '0' || c > '9')
+			break;
+		++numSuffixLength;
+	}
+	outputFilename.resize(outputFilename.size() - numSuffixLength);
+	outputFilename.append(".clx");
+	return outputFilename;
 }
 
 std::optional<IoError> Run(const Options &options)
@@ -85,23 +125,48 @@ std::optional<IoError> Run(const Options &options)
 	std::optional<std::filesystem::path> outputDirFs;
 	if (options.outputDir.has_value())
 		outputDirFs = *options.outputDir;
-	for (const char *inputPath : options.inputPaths) {
-		std::filesystem::path inputPathFs { inputPath };
+
+	if (options.combine) {
+		std::string outputFilename;
+		if (options.outputFilename.has_value()) {
+			outputFilename = *options.outputFilename;
+		} else {
+			outputFilename = DefaultCombinedFilename(options);
+		}
 		std::string outputPath;
 		if (outputDirFs.has_value()) {
-			outputPath = *outputDirFs / inputPathFs.filename().replace_extension("clx");
+			outputPath = *outputDirFs / outputFilename;
 		} else {
-			outputPath = std::filesystem::path(inputPathFs).replace_extension("clx");
+			outputPath = std::filesystem::path(options.inputPaths[0]).parent_path() / outputFilename;
 		}
-		uintmax_t inputFileSize;
-		uintmax_t outputFileSize;
-		if (std::optional<dvl_gfx::IoError> error = Cl2ToClx(inputPath, outputPath.c_str(), options.widths);
-		    error.has_value()) {
-			error->message.append(": ").append(inputPath);
+		std::optional<dvl_gfx::IoError> error = CombineCl2AsClxSheet(
+		    options.inputPaths.data(), options.inputPaths.size(), outputPath.c_str(), options.widths);
+		if (error.has_value())
 			return error;
-		}
-		if (options.remove) {
-			std::filesystem::remove(inputPathFs);
+		return std::nullopt;
+	} else {
+		for (const char *inputPath : options.inputPaths) {
+			std::filesystem::path inputPathFs { inputPath };
+			std::string outputFilename;
+			if (options.outputFilename.has_value()) {
+				outputFilename = *options.outputFilename;
+			} else {
+				outputFilename = inputPathFs.filename().replace_extension("clx");
+			}
+			std::string outputPath;
+			if (outputDirFs.has_value()) {
+				outputPath = *outputDirFs / outputFilename;
+			} else {
+				outputPath = inputPathFs.parent_path() / outputFilename;
+			}
+			if (std::optional<dvl_gfx::IoError> error = Cl2ToClx(inputPath, outputPath.c_str(), options.widths);
+			    error.has_value()) {
+				error->message.append(": ").append(inputPath);
+				return error;
+			}
+			if (options.remove) {
+				std::filesystem::remove(inputPathFs);
+			}
 		}
 	}
 	return std::nullopt;
